# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateRoomRequest {
  hostName: string;
}

export interface Room {
  id: string;
  hostName: string;
  createdAt: string;
  /** Only returned when creating a room (for the host) */
  password?: string;
}

export interface VerifyPasswordRequest {
  password: string;
}

export interface VerifyPasswordResponse {
  valid: boolean;
  token: string;
}

export interface ErrorResponse {
  error: string;
}