# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: string;
  hostName: string;
  createdAt: Date;
  /** Only returned when creating a room (for the host) */
  password?: string;
}