# lib/api-zod/src/generated/types/verifyPasswordResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyPasswordResponse {
  valid: boolean;
  token: string;
}